Print stack trace moved to separate file
