+++ tests/src/stack_trace_print_on_kill.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace utils {
+
+void install_backtrace_handler_on_kill();
+
+}
